Add files via upload
Built an event management website for an event-managing organization called Soundbugs, Goa. The website is to help event organizers to plan and manage their clients’ events like wedding, musical concerts, corporate seminars, birthday celebrations, theme parties, etc. in an efficient manner. Hence, offering a wide range of services to people. The system can lead to error free, secure, reliable and fast management system. It can assist the user/client as well to concentrate on their other activities rather to concentrate on the record keeping. This system is able to manage booking of various events easily. Basically, the project describes how to manage for good performance and better services for the clients. It has both front-end and back-end, leading to the creation of a dynamic platform that redefines event management.

-> Effortless Booking System: The website features an user friendly interface for intuitive booking system that empowers clients to easily reserve their spots at their desired events.
->Review and Feedback Mechanism: A robust review system was incorporated to allow attendees to share their thoughts and opinions on the events they attended.
->Secure Sign-up/Log-in: With user privacy as a top priority, implemented a secure sign-up and log-in process, safeguarding users' personal information and ensuring a trustful environment.
->Secure Payment System: Implementing the latest security protocols, I integrated a robust payment gateway into the website. Attendees can now make transactions with confidence, knowing that their payment information is protected.